Refactor GameView hand section into dedicated view
